Created a new tlsconfig in exthttp and minor changes in cos.go
